feat: Allow adding root certificate authorities
Harbor and Clair can be deployed with certificates signed by certificate
authorities which are not present in the adapters host's root CA set.
If that's the case, the corresponding clients will fail with well known
error: `certificate signed by unknown authority`. Also it happens
because the clients will not trust self-signed certificates, because
they don’t recognise the signer as a trusted Root CA.

In order to support custom CAs and self-signed certificates this commit
adds support for `SCANNER_TLS_CLIENTCAS` and
`SCANNER_TLS_INSECURE_SKIP_VERIFY` config envs.

Resolves: #2

Signed-off-by: Daniel Pacak <pacak.daniel@gmail.com>